pyproject.toml
[tool.poetry]
name = "tdd-project"
version = "0.0.1"
description = ""
authors = ["cezarcorrea <cezar.correa1234@gmail.com>"]
readme = "README.md"
packages = [{include = "store"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic = ">=2.11.7,<3.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
psycopg2-binary = ">=2.9.9,<3.0.0"
pytest = ">=8.4.1,<9.0.0"
pre-commit = ">=4.2.0,<5.0.0"
psycopg = { version = ">=3.2.9,<4.0.0", extras = ["binary", "pool"] }
httpx = ">=0.28.1,<0.29.0"
asyncpg = ">=0.30.0,<0.31.0"
pytest-asyncio = ">=0.23.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
